app/requirements.txt
Flask==2.3.3
redis==5.0.1
mysql-connector-python==8.3.0